+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to
 - The canonical format (`-Tcanon`) produces output that is stable across
   multiple invocations. #2193
 
+- Whitespace immediately around an `IMG` element in an HTML-like label table
+  cell no longer causes a syntax error. #2821
+
 ## [16.1.0] – 2026-09-03
 
 ### Added